Fix remain recursive issues

+++ src/main/java/com/whitespace/BoardScoringService.java
@@ -9,7 +9,7 @@ public interface BoardScoringService {
      * @return 0 > Integer.MAX score of the board strength. If the score is Integer.MIN | Max than that side
      * won or lost.
      */
-    int scoreBoard(ChessBoard board, Player plLayer);
+    double scoreBoard(ChessBoard board, Player plLayer);
 
-    int scorePieces(List<Piece> myPieces, List<Piece> opponentsPieces);
+    double scorePieces(List<Piece> myPieces, List<Piece> opponentsPieces);
 }
